Reorganize format renderers and remove deprecated CLI flags
- Rename renderer files to match requested naming convention:
  * data.go → data_renderers.go
  * simplelist.go → simplelist_renderer.go
  * html.go → html_renderer.go
  * markdown.go → markdown_renderer.go
  * terminal.go → terminal_renderers.go

- Remove deprecated CLI format flags --minimal and --no-color:
  * Update root.go and show.go command definitions
  * Remove legacy flag handling from format manager
  * Update all tests to use --format flag instead
  * Remove legacy support from RenderOptions and RenderRequest structs

- Fix import cycles by ensuring all renderer files use 'format' package
- Update help text to remove references to deprecated flags
- All tests passing with new format-only system
